+++ src/constants/types.ts
@@ -1,0 +1,8 @@
+export interface Dependency {
+  id: string
+  key: string
+  value: string
+  environment: string
+  created_at: string
+  modified_at: string
+}
